Se hace un cambio en como se escriben las plataformas en el YAML

+++ editor/levelEditor/idmaker.h
@@ -10,7 +10,7 @@ class IDMaker{
 
 private:
     std::map<std::string, uint8_t> backgrounds_id;
-    std::map<std::string, uint8_t> platforms_id;
+    std::map<std::string, int> platforms_id;
     std::map<std::string, uint8_t> weapons_id;
     std::map<std::string, uint8_t> armours_id;
     std::map<std::string, uint8_t> walls_id;
@@ -18,7 +18,7 @@ private:
 public:
     IDMaker();
     uint8_t get_id_background(std::string background_name);
-    uint8_t get_id_platform(std::string platform_name);
+    int get_id_platform(std::string platform_name);
     uint8_t get_id_weapon(std::string weapon_name);
     uint8_t get_id_armour(std::string armour_name);
     uint8_t get_id_wall(std::string wall_name);
